Move CLI code to cli directory, move CUDAWorker to cudaworker.py, expose CUDAWorker at top level
