+++ setup.py
@@ -38,6 +38,6 @@ setup(
     install_requires=open("requirements.txt").read().strip().split("\n"),
     entry_points="""
         [console_scripts]
-        dask-cuda-worker=dask_cuda.dask_cuda_worker:go
+        dask-cuda-worker=dask_cuda.cli.dask_cuda_worker:go
       """,
 )
